# internal/httpapi/openapi.yaml
openapi: 3.0.3
info:
  title: OneOps Configuration Registry API
  version: "1.0.0"
  description: >
    The system of record for OneOps artifacts (Configuration State Model §6).
    Authority is a maintained projection computed from the dependency graph (M3).
servers:
  - url: /
security:
  - bearerAuth: []
paths:
...
  /v1/governance/{id}/ratify:
    post:
      summary: Ratify a configuration object
      operationId: ratifyGovernance
      security: [ { bearerAuth: [] } ]
      parameters:
        - { $ref: '#/components/parameters/GovId' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
        - { $ref: '#/components/parameters/IfMatch' }
      responses:
        '200': { $ref: '#/components/responses/GovernanceResult' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '412': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
  /v1/governance/{id}/approve:
    post:
      summary: Approve a configuration object
      operationId: approveGovernance
      security: [ { bearerAuth: [] } ]
      parameters:
        - { $ref: '#/components/parameters/GovId' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
        - { $ref: '#/components/parameters/IfMatch' }
      responses:
        '200': { $ref: '#/components/responses/GovernanceResult' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '412': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
  /v1/governance/{id}/suspend:
    post:
      summary: Suspend a configuration object
      operationId: suspendGovernance
      security: [ { bearerAuth: [] } ]
      parameters:
        - { $ref: '#/components/parameters/GovId' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
        - { $ref: '#/components/parameters/IfMatch' }
      responses:
        '200': { $ref: '#/components/responses/GovernanceResult' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '412': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
  /v1/governance/{id}/deprecate:
    post:
      summary: Deprecate a configuration object
      operationId: deprecateGovernance
      security: [ { bearerAuth: [] } ]
      parameters:
        - { $ref: '#/components/parameters/GovId' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
        - { $ref: '#/components/parameters/IfMatch' }
      responses:
        '200': { $ref: '#/components/responses/GovernanceResult' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '412': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
  /v1/governance/{id}/withdraw:
    post:
      summary: Withdraw a configuration object
      operationId: withdrawGovernance
      security: [ { bearerAuth: [] } ]
      parameters:
        - { $ref: '#/components/parameters/GovId' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
        - { $ref: '#/components/parameters/IfMatch' }
      responses:
        '200': { $ref: '#/components/responses/GovernanceResult' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '412': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
  /v1/governance/{id}/archive:
    post:
      summary: Archive a configuration object into an archival retention class
      operationId: archiveGovernance
      security: [ { bearerAuth: [] } ]
      parameters:
        - { $ref: '#/components/parameters/GovId' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
        - { $ref: '#/components/parameters/IfMatch' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GovernanceArchiveRequest' }
      responses:
        '200': { $ref: '#/components/responses/GovernanceResult' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '412': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
  /v1/governance/{id}:
    get:
      summary: Current governance state of a configuration object
      operationId: getGovernanceState
      security: [ { bearerAuth: [] } ]
      parameters:
        - { $ref: '#/components/parameters/GovId' }
      responses:
        '200':
          description: Current governance state
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GovernanceStateResponse' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
    delete:
      summary: Delete a working-material configuration object
      description: >
        Deletion is permitted only for working_material with no dependents; the
        engine enforces both. The audit chain for the object is preserved.
      operationId: deleteGovernance
      security: [ { bearerAuth: [] } ]
      parameters:
        - { $ref: '#/components/parameters/GovId' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
        - { $ref: '#/components/parameters/IfMatch' }
      responses:
        '200': { $ref: '#/components/responses/GovernanceResult' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '412': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }

  # Governance — read-only query APIs (state / history / audit / verification).
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    GovId:
      name: id
      in: path
      required: true
      description: Configuration object id (cfg_id).
      schema: { type: string }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Idempotency key; used as the OperationID and drives the deterministic audit event id.
      schema: { type: string }
    IfMatch:
      name: If-Match
      in: header
      required: false
      description: Expected row_version (ETag) for optimistic concurrency; omit to skip the check.
      schema: { type: string }
...
  responses:
    Problem:
      description: RFC 7807 problem detail
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    GovernanceResult:
      description: The operation was applied atomically with its audit event.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/GovernanceResponse' }
  schemas:
    GovernanceArchiveRequest:
      type: object
      required: [target_retention]
      properties:
        target_retention:
          type: string
          description: Archival retention class to set.
          enum: [historical_record, historical_evidence, audit_record, superseded_plan]
    GovernanceState:
      type: object
      properties:
        lifecycle: { type: string }
        retention_class: { type: string }
        authority: { type: string }
    GovernanceAudit:
      type: object
      description: Read-only audit metadata; exposes no internal hashes or event ids.
      properties:
        operation_id: { type: string }
        chain_id: { type: string }
        recorded: { type: boolean }
    GovernanceResponse:
      type: object
      properties:
        operation: { type: string }
        cfg_id: { type: string }
        actor: { type: string }
        occurred_at: { type: string, format: date-time }
        removed: { type: boolean }
        row_version: { type: integer, format: int64 }
        state: { $ref: '#/components/schemas/GovernanceState' }
        audit: { $ref: '#/components/schemas/GovernanceAudit' }
    GovernanceStateResponse:
      type: object
      properties:
        cfg_id: { type: string }
        lifecycle: { type: string }
        retention_class: { type: string }
        retention_policy: { type: string }
        authority: { type: string }
        row_version: { type: integer, format: int64 }
        ratified_by: { type: string }
        review_cycle: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    Problem:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
...